+++ src/commands/ext.ts
@@ -26,6 +26,7 @@ export const command = new Command("ext")
       "ext:configure",
       "ext:update",
       "ext:uninstall",
+      "ext:sdk:install",
     ];
 
     for (const commandName of commandNames) {